refactor: convert InputSwitch specs to typescript
