+++ package.json
@@ -164,7 +164,6 @@
     "jest": "^24.5.0",
     "jest-extended": "^0.11.1",
     "jest-serializer-vue": "^2.0.2",
-    "jest-vue-preprocessor": "^1.5.0",
     "lint-staged": "^9.2.1",
     "mini-css-extract-plugin": "^0.9.0",
     "mock-socket": "^9.0.0",
@@ -188,6 +187,7 @@
     "url-loader": "^2.1.0",
     "vue-html-loader": "^1.2.4",
     "vue-i18n-extract": "^1.0.2",
+    "vue-jest": "^3.0.4",
     "vue-loader": "^15.7.1",
     "vue-style-loader": "^4.1.2",
     "vue-template-compiler": "^2.6.10",
